\name{h2o.randomForest.VA}
\alias{h2o.randomForest.VA}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
H2O: Random Forest (ValueArray)
}
\description{
Performs random forest classification using the ValueArray implementation on a data set.
}
\usage{
h2o.randomForest.VA(x, y, data, ntree = 50, depth = 50, sample.rate = 2/3, classwt = NULL, nbins = 100, seed = -1, use_non_local = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
  A vector containing the names or indices of the predictor variables to use in building the random forest model.
  }
  \item{y}{
  The name or index of the response variable. If the data does not contain a header, this is the column index, designated by increasing numbers from left to right. (The response must be either an integer or a categorical variable).
  }
  \item{data}{
  An \code{\linkS4class{H2OParsedDataVA}} object containing the variables in the model.
  }
  \item{ntree}{
  (Optional) Number of trees to grow. (Must be a nonnegative integer).
  }
  \item{depth}{
  (Optional) Maximum depth to grow the tree.
  }
  \item{sample.rate}{
  (Optional) Sampling rate for constructing data from which individual trees are grown.
  }
  \item{classwt}{
  (Optional) Priors of the classes. Need not add up to one. If missing, defaults to all weights set at 1.0.
  }
  \item{nbins}{
  (Optional) Build a histogram of this many bins, then split at best point.
  }
  \item{seed}{
  (Optional) Seed for building the random forest. If \code{seed = -1}, one will automatically be generated by H2O.
  }
  \item{use_non_local}{
  (Optional) A logical value indicating whether to use non-local data for building the model.
  }
}
\details{
This method runs single-node random forest and requires the user import data with \code{\link{h2o.importFile.VA}}, \code{\link{h2o.importFolder.VA}} or one of its variants. Currently, only classification models are supported.
}
\value{
An object of class \code{\linkS4class{H2ORFModelVA}} with slots key, data, and model, where the last is a list of the following components:
\item{ntree }{Number of trees grown.}
\item{classification_error }{Classification error rate.}
\item{tree_sum}{A matrix giving the minimum, mean, and maximum of the tree depth and number of leaves.}
\item{confusion }{Confusion matrix of the prediction.}
}
\examples{
# Run an RF model on iris data
library(h2o)
localH2O = h2o.init(ip = "localhost", port = 54321, startH2O = TRUE, silentUpgrade = TRUE, promptUpgrade = FALSE)
irisPath = system.file("extdata", "iris.csv", package="h2oRClient")
iris.hex = h2o.importFile.VA(localH2O, path = irisPath, key = "iris.hex")
h2o.randomForest.VA(y = 5, x = c(2,3,4), data = iris.hex, ntree = 50, depth = 100)

# Run an RF model on covtype data
covPath = system.file("extdata", "covtype.csv", package="h2oRClient")
covtype.hex = h2o.importFile.VA(localH2O, path = covPath, key = "covtype.hex")
h2o.randomForest.VA(y = "Cover_Type", x = setdiff(colnames(covtype.hex), c("Cover_Type", "Aspect", "Hillshade_9am")), data = covtype.hex, ntree = 50, depth = 150)
h2o.shutdown(localH2O)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
